+++ src/directions/response/mod.rs
@@ -30,7 +30,7 @@ use serde::{Deserialize, Serialize};
 
 /// Directions responses contain the following root elements.
 
-#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Response {
     /// Contains an array of available travel modes. This field is returned when
     /// a request specifies a travel `mode` and gets no results. The array